Refactoring - RepositoryBase
