+++ BlockLoader/DataLayer/BlockRepository.cs
@@ -1,47 +1,34 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.IO;
 using System.Linq;
 using System.Xml.Linq;
 
 namespace BlockLoader.DataLayer
 {
-	public class BlockRepository : IBlockRepository
+	public class BlockRepository : RepositoryBase<Block>, IBlockRepository
 	{
 		private const string BlockElementName = "block";
 		private const string CodeElementName = "code";
 		private const string FootageElementName = "footage";
 		private const string ProgramElementName = "program";
-		private readonly XmlLoader _loader;
-		private readonly string _filePath;
 
-		public BlockRepository(XmlLoader loader, string filePath)
+		public BlockRepository(XmlLoader loader, string filePath) : base(loader, filePath)
 		{
-			_loader = loader;
-			_filePath = filePath;
+		}
+
+		protected override string ElementName
+		{
+			get { return BlockElementName; }
 		}
 
 		public IEnumerable<Block> LoadBlocks()
 		{
-			if (!File.Exists(_filePath))
-			{
-				throw new FileNotFoundException(_filePath);
-			}
-
-			var doc = LoadDocument(_filePath);
-			if (doc == null || doc.Root == null)
-			{
-				throw new InvalidOperationException("Xml file is empty, or invalid.");
-			}
-
-			var blockElements = doc.Root.Elements(BlockElementName);
-
-			return blockElements.Select(CreateBlockFromElement).Where(b => b != null);
+			return Load();
 		}
 
-		private Block CreateBlockFromElement(XElement blockElement)
+		protected override Block Parse(XElement element)
 		{
-			var attributes = blockElement.Attributes().ToDictionary(e => e.Name.LocalName);
+			var attributes = element.Attributes().ToDictionary(e => e.Name.LocalName);
 			if (!attributes.ContainsKey(CodeElementName) || !attributes.ContainsKey(FootageElementName) || !attributes.ContainsKey(ProgramElementName))
 			{
 				throw new InvalidOperationException("Block is missing required attribute.");
@@ -58,12 +45,6 @@ namespace BlockLoader.DataLayer
 			}
 
 			return new Block(code, footageNumber, program);
-
-		}
-
-		private XDocument LoadDocument(string filePath)
-		{
-			return _loader.Load(filePath);
 		}
 	}
 }
